Initial Project Commit
Setup the Flask project and get up and running with a Docker development
environment. Create index and health routes in app.py to ensure that the
server is configured properly and all aspects are running with
docker-compose.

Current setup assumes testing using SQLite3 and will be setup with
seeding data for development. MySQL will be used in production
environment.